# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for user management",
        "title": "User API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/home_details": {
            "get": {
                "description": "Retrieve all home details from MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeDetails"
                ],
                "summary": "Get all home details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HomeDetails.HomeDetails"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new home detail in MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeDetails"
                ],
                "summary": "Create a new home detail",
                "parameters": [
                    {
                        "description": "Home detail object",
                        "name": "homeDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/HomeDetails.HomeDetails"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID of the created home detail",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all home details from MongoDB",
                "tags": [
                    "HomeDetails"
                ],
                "summary": "Delete all home details",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/home_details/search": {
            "get": {
                "description": "Search home details in MongoDB based on a keyword",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeDetails"
                ],
                "summary": "Search home details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HomeDetails.HomeDetails"
                            }
                        }
                    }
                }
            }
        },
        "/home_details/{id}": {
            "get": {
                "description": "Retrieve a home detail by its ID from MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeDetails"
                ],
                "summary": "Get a home detail by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Home detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/HomeDetails.HomeDetails"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing home detail in MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeDetails"
                ],
                "summary": "Update a home detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Home detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated home detail object",
                        "name": "homeDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/HomeDetails.HomeDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Delete a home detail by its ID from MongoDB",
                "tags": [
                    "HomeDetails"
                ],
                "summary": "Delete a home detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Home detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "HomeDetails.HomeDetails": {
            "type": "object",
            "required": [
                "hd_aboutmedesc",
                "hd_aboutmename",
                "hd_cv",
                "hd_desc",
                "hd_image",
                "hd_name"
            ],
            "properties": {
                "hd_aboutmedesc": {
                    "type": "string",
                    "example": "I am a software engineer."
                },
                "hd_aboutmename": {
                    "type": "string",
                    "example": "About Me"
                },
                "hd_cv": {
                    "type": "string",
                    "example": "http://example.com/cv.pdf"
                },
                "hd_desc": {
                    "type": "string",
                    "example": "Description of my home details"
                },
                "hd_id": {
                    "type": "string"
                },
                "hd_image": {
                    "type": "string",
                    "example": "http://example.com/image.jpg"
                },
                "hd_name": {
                    "type": "string",
                    "example": "My Home Details"
                }
            }
        },
...
    }
}